requirements.txt
# Core web framework - Python 3.6 compatible versions
FastAPI==0.68.2
uvicorn==0.15.0

# Data validation and models
pydantic==1.9.2
python-dotenv==0.19.2

# HTTP client and server utilities
httpx==0.22.0
requests==2.27.1

# Database (optional for future features)
sqlalchemy==1.3.24
alembic==1.7.7

# Scheduling and background tasks
APScheduler==3.9.1
pytz==2022.7

# Email functionality
email-validator==1.3.1
premailer==3.10.0

# HTML parsing and templating
beautifulsoup4==4.11.2
jinja2==3.0.3

# File operations
aiofiles==0.8.0

# Data analysis and processing
pandas==1.3.5
numpy==1.19.5

# Technical analysis
ta==0.10.2
yfinance==0.1.87

# Visualization (for potential chart generation)
plotly==5.11.0

# Error handling and retries
tenacity==8.1.0

# Testing
pytest==7.2.2
pytest-asyncio==0.20.3
pytest-mock==3.10.0

# AI/LLM integration - Only Google Gemini
# No OpenAI dependency needed

# Logging enhancements
structlog==22.3.0

# CLI interface
click==8.0.4
rich==12.6.0